// frontend/src/api/model/documentRevisionResponse.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * MapStack API
 * AI学習プラットフォーム MapStack のバックエンドAPI
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentRevisionResponseChangeSummary } from './documentRevisionResponseChangeSummary';
import type { DocumentRevisionResponseCreatedBy } from './documentRevisionResponseCreatedBy';

/**
 * リビジョン情報レスポンス
 */
export interface DocumentRevisionResponse {
  version: string;
  change_summary?: DocumentRevisionResponseChangeSummary;
  created_at: string;
  id: string;
  document_id: string;
  created_by?: DocumentRevisionResponseCreatedBy;
}